QKeyEvent, QContextMenuEvent and QToolBarChangeEvent cleanup

+++ src/gui/kernel/qevent.cpp
@@ -706,7 +706,7 @@ QKeyEvent::QKeyEvent(Type type, int key, Qt::KeyboardModifiers modifiers, const 
 */
 QKeyEvent::QKeyEvent(Type type, int key, Qt::KeyboardModifiers modifiers,
                      quint32 nativeScanCode, quint32 nativeVirtualKey, quint32 nativeModifiers,
-                     const QString &text, bool autorep, ushort count)
+                     const QString &text, bool autorep, int count)
     : QInputEvent(type, modifiers), txt(text), k(key),
       nScanCode(nativeScanCode), nVirtualKey(nativeVirtualKey), nModifiers(nativeModifiers),
       c(count), autor(autorep)
